Route ChatModel streaming probes through StreamingCapabilityDetector.
The verifier now only performs the live probe. Caching lives on the detector so PromptRunner and LlmService share one place for streaming-capability results.
