+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/spi/support/streaming/StreamingCapabilityDetector.kt
@@ -19,20 +19,22 @@ import com.embabel.agent.core.internal.LlmOperations
 import com.embabel.agent.core.internal.streaming.StreamingLlmOperationsFactory
 import com.embabel.common.ai.model.LlmOptions
 import com.embabel.common.util.loggerFor
+import org.springframework.ai.chat.model.ChatModel
+import java.util.Collections
+import java.util.IdentityHashMap
 import java.util.concurrent.ConcurrentHashMap
 
 /**
- * Utility for detecting and caching streaming capability of LLM services.
+ * Detects and caches streaming capability.
  *
- * Not all LLM implementations provide meaningful streaming support. Some may throw
- * UnsupportedOperationException, return empty Flux, or provide stub implementations.
- *
- * This detector caches the results of streaming capability tests (delegated to
- * [StreamingLlmOperationsFactory.supportsStreaming]) using the model as cache key.
+ * PromptRunner goes through [supportsStreaming] with operations and options, cached by model
+ * name. [com.embabel.agent.spi.LlmService.supportsStreaming] goes through the [ChatModel]
+ * overload, which runs [StreamingCapabilityVerifier] once per instance.
  */
 internal object StreamingCapabilityDetector {
     private val logger = loggerFor<StreamingCapabilityDetector>()
-    private val capabilityCache = ConcurrentHashMap<String, Boolean>()
+    private val byModelName = ConcurrentHashMap<String, Boolean>()
+    private val byChatModel = Collections.synchronizedMap(IdentityHashMap<ChatModel, Boolean>())
 
     private const val CACHE_MISS_LOG_MESSAGE = "Cache miss for {}, testing streaming capability..."
 
@@ -51,9 +53,36 @@ internal object StreamingCapabilityDetector {
 
         // Cache by model (or criteria string)
         val cacheKey = llmOptions.model ?: llmOptions.criteria?.toString() ?: "default"
-        return capabilityCache.computeIfAbsent(cacheKey) {
+        return byModelName.computeIfAbsent(cacheKey) {
             logger.debug(CACHE_MISS_LOG_MESSAGE, cacheKey)
             llmOperations.supportsStreaming(llmOptions)
         }
+    }
+
+    /**
+     * Probes [chatModel] via [StreamingCapabilityVerifier] and caches a definitive answer.
+     *
+     * A successful probe or [UnsupportedOperationException] is remembered. Other failures
+     * (missing key, network) answer false for this call but are not cached.
+     */
+    fun supportsStreaming(chatModel: ChatModel): Boolean {
+        byChatModel[chatModel]?.let { return it }
+
+        return try {
+            StreamingCapabilityVerifier.probe(chatModel)
+            byChatModel[chatModel] = true
+            true
+        } catch (_: UnsupportedOperationException) {
+            byChatModel[chatModel] = false
+            false
+        } catch (_: Exception) {
+            false
+        }
+    }
+
+    /** Clears memoized results. For tests only. */
+    internal fun clearCache() {
+        byModelName.clear()
+        byChatModel.clear()
     }
 }
